+++ packages/connector/src/snap.ts
@@ -28,11 +28,13 @@ import type {
   W3CVerifiableCredential,
 } from '@veramo/core';
 
+import { ProviderStore } from './ProviderStore.js';
 import {
   signVerifiableCredential,
   signVerifiablePresentation,
   validateAndSetCeramicSession,
 } from './utils.js';
+import { ViemClient } from './ViemClient.js';
 
 /**
  * Send a request to the Masca snap
@@ -42,16 +44,21 @@ import {
  * @throws Error - if the request fails
  */
 async function sendSnapMethod<T>(
+  context: Masca,
   request: MascaRPCRequest,
   snapId: string
 ): Promise<T> {
-  return window.ethereum.request({
+  const provider = context.providerStore.getCurrentProvider()?.provider;
+  if (!provider) throw new Error('No provider found');
+  // FIXME: currently the wallet_requestSnaps is not in type of EIP-1193 provider since its not standard
+  // therefore "as unknown as any" is used to bypass the type check
+  return provider.request({
     method: 'wallet_invokeSnap',
     params: {
       snapId,
       request,
     },
-  });
+  } as unknown as any);
 }
 
 /**
@@ -67,6 +74,7 @@ async function queryCredentials(
   await validateAndSetCeramicSession.bind(this)();
 
   return sendSnapMethod(
+    this,
     { method: 'queryCredentials', params: params ?? {} },
     this.snapId
   );
@@ -84,6 +92,7 @@ async function createPresentation(
   await validateAndSetCeramicSession.bind(this)();
 
   const result = await sendSnapMethod<Result<VerifiablePresentation>>(
+    this,
     {
       method: 'createPresentation',
       params,
@@ -100,7 +109,7 @@ async function createPresentation(
   }
 
   const signedResult = ResultObject.success(
-    await signVerifiablePresentation(result.data)
+    await signVerifiablePresentation.bind(this)(result.data)
   );
 
   return signedResult;
@@ -121,6 +130,7 @@ async function saveCredential(
   await validateAndSetCeramicSession.bind(this)();
 
   return sendSnapMethod(
+    this,
     {
       method: 'saveCredential',
       params: {
@@ -146,6 +156,7 @@ async function deleteCredential(
   await validateAndSetCeramicSession.bind(this)();
 
   return sendSnapMethod(
+    this,
     {
       method: 'deleteCredential',
       params: {
@@ -162,7 +173,7 @@ async function deleteCredential(
  * @return Result<string> - DID
  */
 async function getDID(this: Masca): Promise<Result<string>> {
-  return sendSnapMethod({ method: 'getDID' }, this.snapId);
+  return sendSnapMethod(this, { method: 'getDID' }, this.snapId);
 }
 
 /**
@@ -170,7 +181,7 @@ async function getDID(this: Masca): Promise<Result<string>> {
  * @return Result<string> - DID method
  */
 async function getSelectedMethod(this: Masca): Promise<Result<string>> {
-  return sendSnapMethod({ method: 'getSelectedMethod' }, this.snapId);
+  return sendSnapMethod(this, { method: 'getSelectedMethod' }, this.snapId);
 }
 
 /**
@@ -178,7 +189,7 @@ async function getSelectedMethod(this: Masca): Promise<Result<string>> {
  * @return Result<string[]> - list of available DID methods
  */
 async function getAvailableMethods(this: Masca): Promise<Result<string[]>> {
-  return sendSnapMethod({ method: 'getAvailableMethods' }, this.snapId);
+  return sendSnapMethod(this, { method: 'getAvailableMethods' }, this.snapId);
 }
 
 /**
@@ -192,6 +203,7 @@ async function switchDIDMethod(
 ): Promise<Result<AvailableMethods>> {
   if (this.supportedMethods.includes(method)) {
     return sendSnapMethod(
+      this,
       { method: 'switchDIDMethod', params: { didMethod: method } },
       this.snapId
     );
@@ -204,11 +216,11 @@ async function switchDIDMethod(
  * @return Result<boolean> - true if the switch was successful
  */
 async function togglePopups(this: Masca): Promise<Result<boolean>> {
-  return sendSnapMethod({ method: 'togglePopups' }, this.snapId);
+  return sendSnapMethod(this, { method: 'togglePopups' }, this.snapId);
 }
 
 /**
- * Adds origin of the current dapp to the list of friendly dapps. This will disable popups from appearing while using the dapp.
+ * Adds origin of the current dapp to the list of trusted dapps. This will disable popups from appearing while using the dapp.
  *
  * @return Result<boolean> - true if the addition was successful
  */
@@ -217,13 +229,14 @@ async function addTrustedDapp(
   origin: string
 ): Promise<Result<boolean>> {
   return sendSnapMethod(
+    this,
     { method: 'addTrustedDapp', params: { origin } },
     this.snapId
   );
 }
 
 /**
- * Removes origin of the current dapp from the list of friendly dapps. This will enable popups while using the dapp.
+ * Removes origin of the current dapp from the list of trusted dapps. This will enable popups while using the dapp.
  *
  * @return Result<boolean> - true if the removal was successful
  */
@@ -232,6 +245,7 @@ async function removeTrustedDapp(
   origin: string
 ): Promise<Result<boolean>> {
   return sendSnapMethod(
+    this,
     { method: 'removeTrustedDapp', params: { origin } },
     this.snapId
   );
@@ -245,7 +259,7 @@ async function removeTrustedDapp(
 async function getCredentialStore(
   this: Masca
 ): Promise<Result<Record<AvailableCredentialStores, boolean>>> {
-  return sendSnapMethod({ method: 'getCredentialStore' }, this.snapId);
+  return sendSnapMethod(this, { method: 'getCredentialStore' }, this.snapId);
 }
 
 /**
@@ -256,6 +270,7 @@ async function getAvailableCredentialStores(
   this: Masca
 ): Promise<Result<string[]>> {
   return sendSnapMethod(
+    this,
     { method: 'getAvailableCredentialStores' },
     this.snapId
   );
@@ -273,6 +288,7 @@ async function setCredentialStore(
   value: boolean
 ): Promise<Result<boolean>> {
   return sendSnapMethod(
+    this,
     { method: 'setCredentialStore', params: { store, value } },
     this.snapId
   );
@@ -285,7 +301,7 @@ async function setCredentialStore(
 async function getAccountSettings(
   this: Masca
 ): Promise<Result<MascaAccountConfig>> {
-  return sendSnapMethod({ method: 'getAccountSettings' }, this.snapId);
+  return sendSnapMethod(this, { method: 'getAccountSettings' }, this.snapId);
 }
 
 /**
@@ -293,7 +309,7 @@ async function getAccountSettings(
  * @return Result<MascaConfig> - Masca settings
  */
 async function getSnapSettings(this: Masca): Promise<Result<MascaConfig>> {
-  return sendSnapMethod({ method: 'getSnapSettings' }, this.snapId);
+  return sendSnapMethod(this, { method: 'getSnapSettings' }, this.snapId);
 }
 
 /**
@@ -305,7 +321,11 @@ async function resolveDID(
   this: Masca,
   did: string
 ): Promise<Result<DIDResolutionResult>> {
-  return sendSnapMethod({ method: 'resolveDID', params: { did } }, this.snapId);
+  return sendSnapMethod(
+    this,
+    { method: 'resolveDID', params: { did } },
+    this.snapId
+  );
 }
 
 /**
@@ -320,6 +340,7 @@ async function createCredential(
   await validateAndSetCeramicSession.bind(this)();
 
   const result = await sendSnapMethod(
+    this,
     {
       method: 'createCredential',
       params,
@@ -355,6 +376,7 @@ async function setCurrentAccount(
   params: SetCurrentAccountRequestParams
 ): Promise<Result<boolean>> {
   return sendSnapMethod(
+    this,
     {
       method: 'setCurrentAccount',
       params,
@@ -374,6 +396,7 @@ async function verifyData(
   params: VerifyDataRequestParams
 ): Promise<Result<boolean | IVerifyResult>> {
   return sendSnapMethod(
+    this,
     {
       method: 'verifyData',
       params,
@@ -393,6 +416,7 @@ async function handleCredentialOffer(
   params: HandleCredentialOfferRequestParams
 ): Promise<Result<VerifiableCredential[]>> {
   return sendSnapMethod(
+    this,
     {
       method: 'handleCredentialOffer',
       params,
@@ -412,6 +436,7 @@ async function handleAuthorizationRequest(
   params: HandleAuthorizationRequestParams
 ): Promise<Result<void>> {
   return sendSnapMethod(
+    this,
     {
       method: 'handleAuthorizationRequest',
       params,
@@ -431,6 +456,7 @@ async function setCeramicSession(
   serializedSession: string
 ): Promise<Result<boolean>> {
   return sendSnapMethod(
+    this,
     {
       method: 'setCeramicSession',
       params: { serializedSession },
@@ -448,6 +474,7 @@ async function validateStoredCeramicSession(
   this: Masca
 ): Promise<Result<boolean>> {
   return sendSnapMethod(
+    this,
     {
       method: 'validateStoredCeramicSession',
     },
@@ -462,6 +489,7 @@ async function validateStoredCeramicSession(
  */
 async function exportStateBackup(this: Masca): Promise<Result<string>> {
   return sendSnapMethod(
+    this,
     {
       method: 'exportStateBackup',
     },
@@ -480,6 +508,7 @@ async function importStateBackup(
   params: ImportStateBackupRequestParams
 ): Promise<Result<boolean>> {
   return sendSnapMethod(
+    this,
     {
       method: 'importStateBackup',
       params,
@@ -495,6 +524,7 @@ async function importStateBackup(
  */
 async function getWalletId(this: Masca): Promise<Result<string>> {
   return sendSnapMethod(
+    this,
     {
       method: 'getWalletId',
     },
@@ -512,6 +542,7 @@ async function signData(
   params: SignDataRequestParams
 ): Promise<Result<string>> {
   return sendSnapMethod(
+    this,
     {
       method: 'signData',
       params,
@@ -535,9 +566,17 @@ export class Masca {
 
   public readonly supportedMethods: AvailableMethods[];
 
+  public providerStore: ProviderStore;
+
+  public viemClient: ViemClient;
+
   public constructor(snapId: string, supportedMethods: AvailableMethods[]) {
     this.snapId = snapId;
     this.supportedMethods = supportedMethods;
+    this.providerStore = new ProviderStore();
+    this.viemClient = new ViemClient({
+      provider: this.providerStore.getCurrentProvider()?.provider,
+    });
   }
 
   public getMascaApi = (): MascaApi => ({