+++ packages/dapp/src/app/[locale]/app/(public)/share-presentation/[id]/formatedView.tsx
@@ -4,12 +4,14 @@ import { useMemo } from 'react';
 import { usePathname, useRouter } from 'next/navigation';
 import {
   CheckCircleIcon,
+  DocumentDuplicateIcon,
   ExclamationCircleIcon,
 } from '@heroicons/react/24/outline';
 import { Pagination, Tooltip } from '@nextui-org/react';
 import { VerifiableCredential } from '@veramo/core';
 import { useTranslations } from 'next-intl';
 
+import { copyToClipboard } from '@/utils/string';
 import CredentialPanel from './credentialPanel';
 
 export const FormatedView = ({
@@ -40,7 +42,7 @@ export const FormatedView = ({
   return (
     <>
       <div className="dark:bg-navy-blue-800 h-full w-full rounded-3xl bg-white shadow-lg">
-        <div className="dark:from-navy-blue-700 dark:to-navy-blue-700 flex max-w-full flex-col-reverse items-center space-x-4 rounded-t-2xl bg-gradient-to-br from-pink-100 to-orange-100 px-10 pt-6 sm:flex-row">
+        <div className="dark:from-navy-blue-700 dark:to-navy-blue-700 flex max-w-full flex-col-reverse items-center space-x-4 rounded-t-2xl bg-gradient-to-br from-pink-100 to-orange-100 px-10 pb-2 pt-6 sm:flex-row">
           <div className="flex w-full">
             <div className="flex flex-col space-y-4">
               <div className="flex flex-col">
@@ -48,8 +50,36 @@ export const FormatedView = ({
                   {t('holder')}
                 </h2>
                 <h1 className="font-ubuntu dark:text-orange-accent-dark text-left text-lg font-medium text-pink-500 sm:text-xl md:text-2xl lg:truncate">
-                  {holder.substring(0, 20)}...
-                  {holder.substring(holder.length, holder.length - 10)}
+                  <div className="mt-2 flex items-center">
+                    <Tooltip
+                      content={holder}
+                      className="border-navy-blue-300 bg-navy-blue-100 text-navy-blue-700"
+                    >
+                      <a
+                        href={`https://dev.uniresolver.io/#${holder}`}
+                        target="_blank"
+                        rel="noreferrer"
+                        className="font-ubuntu dark:text-orange-accent-dark text-left text-lg font-medium text-pink-500 underline sm:text-xl md:text-2xl lg:truncate"
+                      >
+                        {`${holder.substring(
+                          0,
+                          holder.lastIndexOf(':')
+                        )}:${holder
+                          .split(':')
+                          [holder.split(':').length - 1].slice(
+                            0,
+                            6
+                          )}...${holder.slice(-4)}`}
+                      </a>
+                    </Tooltip>
+                    <button
+                      onClick={() => {
+                        copyToClipboard(holder);
+                      }}
+                    >
+                      <DocumentDuplicateIcon className="animated-transition dark:text-orange-accent-dark ml-1 h-5 w-5 text-pink-500 hover:opacity-80" />
+                    </button>
+                  </div>
                 </h1>
               </div>
               {issuanceDate && (
@@ -91,7 +121,7 @@ export const FormatedView = ({
           </div>
         </div>
         <div className="px-4 pb-6">
-          <div className="flex justify-start px-2">
+          <div className="ml-[14px] flex justify-start px-2">
             <Pagination
               loop
               color="success"