+++ packages/dapp/src/app/api/veramoSetup.ts
@@ -1,5 +1,4 @@
 import { getDidKeyResolver as didKeyResolver } from '@blockchain-lab-um/did-provider-key';
-import { JsonRpcProvider } from '@ethersproject/providers';
 import {
   createAgent,
   type ICredentialVerifier,
@@ -16,6 +15,8 @@ import {
 import { getDidPkhResolver as didPkhResolver } from '@veramo/did-provider-pkh';
 import { DIDResolverPlugin } from '@veramo/did-resolver';
 import { Resolver } from 'did-resolver';
+import { getResolver as didEnsResolver } from 'ens-did-resolver';
+import { JsonRpcProvider } from 'ethers';
 import { getResolver as didEthrResolver } from 'ethr-did-resolver';
 
 export type Agent = TAgent<
@@ -26,13 +27,13 @@ const networks = [
   {
     name: 'mainnet',
     provider: new JsonRpcProvider(
-      'https://mainnet.infura.io/v3/bf246ad3028f42318f2e996a7aa85bfc'
+      process.env.MAINNET_RPC_URL || 'https://mainnet.infura.io/v3/'
     ),
   },
   {
     name: 'sepolia',
     provider: new JsonRpcProvider(
-      'https://sepolia.infura.io/v3/bf246ad3028f42318f2e996a7aa85bfc'
+      process.env.SEPOLIA_RPC_URL || 'https://sepolia.infura.io/v3/'
     ),
     chainId: '0xaa36a7',
   },
@@ -55,6 +56,9 @@ export const getAgent = async (): Promise<Agent> => {
           ...didEthrResolver({ networks }),
           ...didPkhResolver(),
           ...didKeyResolver(),
+          ...didEnsResolver({
+            rpcUrl: process.env.MAINNET_RPC_URL,
+          }),
         }),
       }),
     ],