+++ packages/dapp/src/components/ProjectIcon/index.tsx
@@ -9,7 +9,7 @@ export interface ProjectIconProps {
 
 const ProjectIcon = ({ icon, href, alt }: ProjectIconProps) => (
   <div className="relative h-24 w-full sm:w-36">
-    <Link href={href}>
+    <Link href={href} target="_blank">
       <Image src={icon} alt={alt} fill objectFit="contain" />
     </Link>
   </div>