+++ packages/dapp/src/config/wagmiConfig.ts
@@ -1,0 +1,15 @@
+import { createConfig, http } from 'wagmi';
+import { mainnet, polygon, polygonMumbai, sepolia } from 'wagmi/chains';
+import { injected } from 'wagmi/connectors';
+
+export const config = createConfig({
+  chains: [mainnet, sepolia, polygon, polygonMumbai],
+  connectors: [injected({ target: 'metaMask' })],
+  ssr: true,
+  transports: {
+    [mainnet.id]: http(),
+    [sepolia.id]: http(),
+    [polygon.id]: http(),
+    [polygonMumbai.id]: http(),
+  },
+});