+++ packages/dapp/src/utils/networks.ts
@@ -1,7 +1,6 @@
 export const NETWORKS: Record<string, string> = {
   '0x1': 'Ethereum',
-  '0xaa36a7': 'Sepolia Testnet',
-  '0x5': 'Goerli Testnet',
+  '0xaa36a7': 'Sepolia',
   '0x89': 'Polygon',
   '0x13881': 'Polygon Mumbai',
 };
@@ -9,8 +8,9 @@ export const NETWORKS: Record<string, string> = {
 export const NETWORKS_BY_DID: Record<string, string[]> = {
   'did:ethr': ['*'],
   'did:pkh': ['0x1', '0x89'],
-  'did:polygonid': ['0x1', '0x5', '0x89', '0x13881'],
-  'did:iden3': ['0x1', '0x5', '0x89', '0x13881'],
+  'did:ens': ['0x1'],
+  'did:polygonid': ['0x1', '0x89', '0x13881'],
+  'did:iden3': ['0x1', '0x89', '0x13881'],
 };
 
 export function getAvailableNetworksList(method: string): string[] {